requirements.txt
Django>=4.0
python-telegram-bot>=20.0
openai>=1.0.0
python-dotenv
google-generativeai>=0.8.5